refactor : 유저 업데이트 작업을 진행하는 클래스 이름 변경
